+++ lib/screens/editor/addr_chooser.dart
@@ -1,6 +1,6 @@
 import 'package:every_door/constants.dart';
 import 'package:every_door/models/address.dart';
-import 'package:every_door/providers/geolocation.dart';
+import 'package:every_door/providers/location.dart';
 import 'package:every_door/providers/imagery.dart';
 import 'package:every_door/providers/osm_data.dart';
 import 'package:flutter/material.dart';